+++ example/files/load.dart
@@ -1,0 +1,13 @@
+import 'package:objd/core.dart';
+
+class LoadFile extends Widget {
+  /// The load file is put in a seperate widget to split the code in several files
+  LoadFile();
+
+  @override
+  Widget generate(Context context) {
+    return For.of([
+      // put your load widgets here
+    ]);
+  }
+}