+++ example/files/main.dart
@@ -1,0 +1,14 @@
+import 'package:objd/core.dart';
+
+class MainFile extends Widget {
+  /// The main file is put in a seperate widget to split the code in several files
+  MainFile();
+
+  @override
+  Widget generate(Context context) {
+    return For.of([
+      // put your tick widgets here
+      Particle()
+    ]);
+  }
+}